Move lifecycle list to @netlify/config
